Extract svg-inclusion methods into support module
