+++ lib/jekyll-plugin-platoniq-journal.rb
@@ -2,6 +2,7 @@
 
 require "jekyll"
 require "kramdown"
+require "support/includes_file"
 require "jekyll-plugin-platoniq-journal/call_to_action_block_tag"
 require "jekyll-plugin-platoniq-journal/numbers_tag"
 require "jekyll-plugin-platoniq-journal/quote_tag"
